+++ src/Compute/Compute/ChangeLog.md
@@ -20,6 +20,7 @@
 
 -->
 ## Upcoming Release
+* Added optional parameter `-OrchestrationMode` to `New-AzVmss` and `New-AzVmssConfig`
 * Updated the following cmdlets to work when the resource uses a remote image source using AKS or Shared Image Gallery.
     - `Update-AzVm`
     - `Update-AzVmss`